+++ webroot/index.php
@@ -1,6 +1,8 @@
 <?php
 
 require_once __DIR__ . "/../resources/autoload.php";
+
+
 require getTemplatePath("header.php");
 require getTemplatePath("home.php");
 require getTemplatePath("footer.php");