+++ CHANGELOG.md
@@ -7,6 +7,9 @@ For details on the changes in each release, see [the Releases page](https://gith
 ### 1.7 -> 1.8
 - the [webhook] section of the config file can be removed
 - all mail templates are no longer PHP files, now they are `.html.twig`
+- the `mail_overrides` directory has been renamed to `mail`
+- the `templates_overrides` directory has been renamed to `templates`
+- the `overrides` directory has been renamed to `domain_overrides`
 
 ### 1.6 -> 1.7